Add tests for array of array and 2-dimensional array
